requirements.txt
alembic~=1.16.2
pytest~=8.4.1
pydantic~=2.11.7
sqlalchemy~=2.0.41
fastapi~=0.115.14
psycopg2~=2.9.10
uvicorn~=0.35.0
starlette~=0.46.2
pydantic-settings~=2.10.1
python-jose~=3.5.0